#209 Shrink BlueMapAPIConnector's marker-mutation surface
